refactor(rolldown_plugin_json): move utils logic to utils.rs
